requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
pillow>=10.0.0
huggingface_hub>=0.19.0
accelerate>=0.25.0
tensorboard>=2.14.0
